requirements.txt
flask
streamlit
requests
google-generativeai
python-dotenv